+++ .atomist/editors/AddTypeScriptEditor.ts
@@ -44,8 +44,9 @@ export class AddTypeScriptEditor implements EditProject {
     public edit(project: Project) {
 
         if (!IsRugArchive(project)) {
-            console.log("This project does not appear to be a Rug project");
-            return;
+            const err = "project does not appear to be a Rug project";
+            console.log(err);
+            throw new Error(err);
         }
 
         project.editWith("AddTypeScript", {});