+++ .atomist/editors/AddTypeScriptEventHandler.ts
@@ -53,8 +53,9 @@ export class AddTypeScriptEventHandler implements EditProject {
 
     public edit(project: Project) {
         if (!IsRugArchive(project)) {
-            console.log("This project does not appear to be a Rug archive project");
-            return;
+            const err = "project does not appear to be a Rug project";
+            console.log(err);
+            throw new Error(err);
         }
 
         project.editWith("AddTypeScript", {});