+++ .atomist/editors/AddTypeScriptGenerator.ts
@@ -43,8 +43,9 @@ export class AddTypeScriptGenerator implements EditProject {
 
     public edit(project: Project) {
         if (!IsRugArchive(project)) {
-            console.log("This project does not appear to be a Rug archive project");
-            return;
+            const err = "project does not appear to be a Rug project";
+            console.log(err);
+            throw new Error(err);
         }
 
         project.editWith("AddTypeScript", {});